test: move test cases to testing lib for Tag (#35783)
* test: move test cases to testing lib for Tag

* chore: improve code

* fix: use react-dom/test-utils to make asserts working for click event

* chore: udpate

* chore: remove useless file

* test: diff

* chore: rename to .tsx
